requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
Django==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.12.1
numpy==1.19.1
pandas==1.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wincertstore==0.2
